% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicos.R
\name{basicos}
\alias{basicos}
\title{basicos}
\usage{
basicos(url)
}
\arguments{
\item{url}{url a los datos del Ministerio del Interior.}
}
\value{
Dataframe con los datos globales de las mesas.
}
\description{
Esta función descarga los datos globales de las mesas de unas elecciones, los importa al espacio de trabajo, y los formatea.
}
