% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatos.R
\name{candidatos}
\alias{candidatos}
\title{candidatos}
\usage{
candidatos(tipo_eleccion, anno, mes)
}
\arguments{
\item{tipo_eleccion}{El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".}

\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}
}
\value{
Dataframe con los datos de candidatos.
}
\description{
Esta función descarga los datos de los candidatos de las listas electorales de las elecciones seleccionadas, los formatea, y los importa al espacio de trabajo.
}
