% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatos_nosenado.R
\name{candidatos_nosenado}
\alias{candidatos_nosenado}
\title{candidatos_nosenado}
\usage{
candidatos_nosenado(tipo, anno, mes)
}
\arguments{
\item{tipo}{Code for the type of election.}

\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}
}
\value{
data.frame with the data of candidates.
}
\description{
`candidatos_nosenado()` downloads, formats and imports to the environment the data of the candidates from the electoral lists of the selected elections.
}
\keyword{internal}
