% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatos_senado.R
\name{candidatos_senado}
\alias{candidatos_senado}
\title{candidatos_senado}
\usage{
candidatos_senado(anno, mes, nivel)
}
\arguments{
\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}

\item{nivel}{The administrative level for which the data is wanted ("mesa" for polling stations or "municipio" for municipalities).}
}
\value{
data.frame with the data for the Senate candidates.
}
\description{
`candidatos_senado()` downloads, formats and imports to the environment the data of the Senate candidates of the selected elections.
}
\keyword{internal}
