% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidaturas.R
\name{candidaturas}
\alias{candidaturas}
\title{candidaturas}
\usage{
candidaturas(url)
}
\arguments{
\item{url}{url a los datos del Ministerio del Interior de una elección determinada.}
}
\value{
Dataframe con los datos de candidaturas.
}
\description{
Esta función descarga los datos de candidaturas de unas elecciones, los importa al espacio de trabajo, y los formatea.
}
