% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Cleans up the files and directories created during the process}
\usage{
cleanup(dirs)
}
\arguments{
\item{dirs}{The directories to remove.}
}
\description{
Cleans up the files and directories created during the process
}
\keyword{internal}
