% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_ccaa}
\alias{codigos_ccaa}
\title{Codigos de las comunidades autónomas de España.}
\format{
Un conjunto de datos con 19 filas y 3 variables:
\describe{
  \item{codigo_ccaa}{código infoelectoral de la comunidad autónoma}
  \item{codigo_ccaa_ine}{código INE de la comunidad autónoma}
  \item{provincia}{nombre oficial de la comunidad autónoma}
}
}
\source{
\url{http://www.infoelectoral.mir.es/infoelectoral/min/areaDescarga.html?method=inicio}
}
\usage{
codigos_ccaa
}
\description{
Este conjunto de datos contiene los códigos de las comunidades autónomas tal y como aparecen en el diseño de registro de infoelectoral, los códigos INE y los nombres de las comunidades autónomas españolas.
}
\keyword{datasets}
