% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_partidos}
\alias{codigos_partidos}
\title{Recoded party names}
\format{
A dataset with the names of the electoral party or coalition
\describe{
  \item{anno}{Year of the election}
  \item{mes}{Month of the election}
  \item{codigo_partido_nacional}{The national accumulation code for the electoral party or coalition}
  \item{partido}{The recoded name for the electoral party or coalition.}
}
}
\usage{
codigos_partidos
}
\description{
This dataset contains a list of recoded electoral party or coalition names with their correspondent national codes. For example: 'PSOE' when the original name is 'PSA-PSOE', 'PSOE-PROGR.' or 'PSOE-A'. This recodification helps the longitudinal analysis of the electoral results, avoiding the many variations in the party and coalition names.
}
\keyword{datasets}
