% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_provincias}
\alias{codigos_provincias}
\title{Administrative codes for spanish provinces.}
\format{
A dataset with 52 rows and 5 columns:
\describe{
  \item{codigo_ccaa}{Code given to the autonomous communities by the Ministry of Interior (not the same as the National Institute of Statistics codes)}
  \item{codigo_ccaa_ine}{Code given to the autonomous communities by the National Institute of Statistics}
  \item{ccaa}{Official name of the autonomous communities}
  \item{codigo_provincia}{Code given to the provinces by the National Institute of Statistics}
  \item{provincia}{Official name of the provinces}
}
}
\source{
\url{https://www.ine.es/daco/daco42/codmun/cod_provincia.htm}
}
\usage{
codigos_provincias
}
\description{
This dataset contains the National Institute of Statistics administrative codes for spanish provinces with their official names.
}
\keyword{datasets}
