% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bin.R
\name{download_bin}
\alias{download_bin}
\title{Downloads the elections zip file}
\usage{
download_bin(url, tempfile)
}
\arguments{
\item{url}{The URL of the elections zip to download}

\item{tempfile}{The path to save the downloaded file}
}
\description{
Downloads the elections zip file
}
\keyword{internal}
