% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mesas.R
\name{download_mesas}
\alias{download_mesas}
\title{download_mesas}
\usage{
download_mesas(tipoeleccion, yr, mes, dir)
}
\arguments{
\item{tipoeleccion}{El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".}

\item{yr}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}

\item{dir}{La ruta a la carpeta donde se quiere guardar el output.}
}
\value{
Dataframe con los datos de voto a candidaturas por mesas.
}
\description{
Esta función descarga los datos de voto a candidaturas a nivel de mesas de las elecciones seleccionadas, los formatea y los guarda en el directorio especificado.
}
