% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_type_code.R
\name{election_type_code}
\alias{election_type_code}
\title{Generates the election type code from a string}
\usage{
election_type_code(tipo_eleccion)
}
\arguments{
\item{tipo_eleccion}{The type of choice you want to download. The accepted values are "congreso", "senado", "europeas" o "municipales".}
}
\value{
A string
}
\description{
Generates the election type code from a string
}
\keyword{internal}
