% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fechas_elecciones}
\alias{fechas_elecciones}
\title{Election dates}
\format{
A dataset with the dates of the elections
\describe{
  \item{tipo_eleccion}{Election level (local, general or european parliament)}
  \item{anno}{Year of the election}
  \item{mes}{Month of the election}
  \item{day}{Day of the election}
}
}
\source{
\url{https://www.juntaelectoralcentral.es/cs/jec/elecciones}
}
\usage{
fechas_elecciones
}
\description{
This dataset contains the dates of the local, general and european elections in Spain from 1977 to 2023.
}
\keyword{datasets}
