% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_url.R
\name{generate_url}
\alias{generate_url}
\title{Generates the downloads URL of the election zip file}
\usage{
generate_url(tipo, anno, mes, level)
}
\arguments{
\item{tipo}{The code of the type of election you want to download.}

\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}
}
\value{
A string
}
\description{
Generates the downloads URL of the election zip file
}
\keyword{internal}
