% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni.R
\name{municipios}
\alias{municipios}
\title{municipios}
\usage{
municipios(tipoeleccion, yr, mes)
}
\arguments{
\item{tipoeleccion}{El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".}

\item{yr}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}
}
\value{
Dataframe con los datos de voto a candidaturas por mesas.
}
\description{
Esta función descarga los datos de voto a candidaturas a nivel de municipio de las elecciones seleccionadas, los importa al espacio de trabajo, y los formatea.
}
