% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provincias}
\alias{provincias}
\title{Codigos INE de las provincias de Espa침a.}
\format{
Un conjunto de datos con 52 filas y 2 variables:
\describe{
  \item{codigo_provincia}{c칩digo INE de la provincia}
  \item{provincia}{Nombre oficial de la provincia}
}
}
\source{
\url{https://www.ine.es/daco/daco42/codmun/cod_provincia.htm}
}
\usage{
provincias
}
\description{
Este conjunto de datos contiene los c칩digos INE y los nombres de las provincias espa침olas.
}
\keyword{datasets}
