% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read03.R
\name{read03}
\alias{read03}
\title{Reads the candidacy files}
\usage{
read03(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 03.}
}
\value{
data.frame
}
\description{
Reads the candidacy files
}
\keyword{internal}
