% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read04.R
\name{read04}
\alias{read04}
\title{Reads the candidates info files}
\usage{
read04(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 04.}
}
\value{
data.frame
}
\description{
Reads the candidates info files
}
\keyword{internal}
