% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read05.R
\name{read05}
\alias{read05}
\title{Lee los ficheros de datos comunes de municipios}
\usage{
read05(file, tempd)
}
\arguments{
\item{file}{Ruta al fichero .DAT que comienza con 05 y que contienen los datos comunes de municipios}
}
\value{
data.frame
}
\description{
Lee los ficheros de datos comunes de municipios
}
\keyword{internal}
