% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read06.R
\name{read06}
\alias{read06}
\title{Reads the candidacy info at municipality level files}
\usage{
read06(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 06.

 @return data.frame}
}
\description{
Reads the candidacy info at municipality level files
}
\keyword{internal}
