% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read07.R
\name{read07}
\alias{read07}
\title{Reads the basic data files for the area superior to the municipality (provinces, islands, etc...)}
\usage{
read07(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 07.}
}
\value{
data.frame
}
\description{
Reads the basic data files for the area superior to the municipality (provinces, islands, etc...)
}
\keyword{internal}
