% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read08.R
\name{read08}
\alias{read08}
\title{Reads the candidacy files at the area superior to the municipality (provinces, islands, etc...)}
\usage{
read08(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 08.}
}
\value{
data.frame
}
\description{
Reads the candidacy files at the area superior to the municipality (provinces, islands, etc...)
}
\keyword{internal}
