% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read09.R
\name{read09}
\alias{read09}
\title{Lee los ficheros de datos comunes de mesas}
\usage{
read09(file, tempd)
}
\arguments{
\item{file}{Ruta al fichero .DAT que comienza con 09 y que contienen los datos comunes de mesas y del CERA}
}
\value{
data.frame
}
\description{
Lee los ficheros de datos comunes de mesas
}
\keyword{internal}
