% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10.R
\name{read10}
\alias{read10}
\title{Reads the candidacy info at polling station level files}
\usage{
read10(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 10.}
}
\value{
data.frame
}
\description{
Reads the candidacy info at polling station level files
}
\keyword{internal}
