% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read11.R
\name{read11}
\alias{read11}
\title{Reads the basic data files for municipalities of 250 inhabitants or less}
\usage{
read11(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 11.

 @return data.frame}
}
\description{
Reads the basic data files for municipalities of 250 inhabitants or less
}
\keyword{internal}
