% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read12.R
\name{read12}
\alias{read12}
\title{Reads the candidacy files for small municipalities}
\usage{
read12(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 12.}
}
\value{
data.frame
}
\description{
Reads the candidacy files for small municipalities
}
\keyword{internal}
