% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renta}
\alias{renta}
\title{Mean income at the census section level (INE)}
\format{
A dataset with more than 34.000 rows and 2 columns:
\describe{
  \item{codigo_seccion}{Code given to the census section made by the combination of the codes of the province, the municipality, the district and the section.}
  \item{renta}{Mean income of the census section in euros}
}
}
\source{
\url{https://www.ine.es/experimental/atlas/experimental_atlas.htm}
}
\usage{
renta
}
\description{
This dataset contains the mean income of each census section
}
\keyword{datasets}
