% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bin.R
\name{safe_download_infoelectoral}
\alias{safe_download_infoelectoral}
\title{Downloads the elections zip file}
\usage{
safe_download_infoelectoral(url, tempfile, get = httr::GET)
}
\arguments{
\item{url}{The URL of the elections zip to download}

\item{tempfile}{The path to save the downloaded file}
}
\value{
The path to the downloaded file, or \code{NULL} if the remote
  resource is unavailable.
}
\description{
Downloads the elections zip file
}
\keyword{internal}
