% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senado.R
\name{senado}
\alias{senado}
\title{senado}
\usage{
senado(anno, mes, nivel)
}
\arguments{
\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}

\item{nivel}{El nivel de desagregación para el que se quieren los datos. Puede ser "mesa" o "municipio".}
}
\value{
Dataframe con los datos de los senadores.
}
\description{
Esta función descarga los datos de los candidatos al Senado a nivel de mesa o municipio.
}
