% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senado_mesas.R
\name{senado_mesas}
\alias{senado_mesas}
\title{senado_mesas}
\usage{
senado_mesas(anno, mes)
}
\arguments{
\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}
}
\value{
data.frame with the data for the Senate candidates.
}
\description{
`senado_mesas` downloads the Senate candidates data at the polling station level.
}
\keyword{internal}
