% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senado_municipios.R
\name{senado_municipios}
\alias{senado_municipios}
\title{senado_municipios}
\usage{
senado_municipios(anno, mes)
}
\arguments{
\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}
}
\value{
data.frame with the data for the Senate candidates.
}
\description{
`senado_mesas` downloads the Senate candidates data at the municipality level.
}
\keyword{internal}
