% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidatos_nosenado.R
\name{candidatos_nosenado}
\alias{candidatos_nosenado}
\title{Descarga datos de candidatos (no Senado)}
\usage{
candidatos_nosenado(tipo, anno, mes)
}
\arguments{
\item{tipo}{Código del tipo de proceso electoral}

\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}
}
\value{
Dataframe con los datos de candidatos.
}
\description{
Esta función descarga los datos de los candidatos de las listas electorales de las elecciones seleccionadas, los formatea, y los importa al espacio de trabajo.
}
