% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_ccaa}
\alias{codigos_ccaa}
\title{Administrative codes for spanish autonomous communities.}
\format{
A dataset with 19 rows and 3 columns:
\describe{
  \item{codigo_ccaa}{Code given to the autonomous communities by the Ministry of Interior (not the same as the National Institute of Statistics codes)}
  \item{codigo_ccaa_ine}{Code given to the autonomous communities by the National Institute of Statistics}
  \item{provincia}{Official name of the provinces}
}
}
\source{
\url{https://infoelectoral.interior.gob.es/opencms/es/elecciones-celebradas/area-de-descargas/}
}
\usage{
codigos_ccaa
}
\description{
This dataset contains the codes given to the autonomus communities by the Ministry of Interior and the ones given by the National Institute of Statistics with their official names.
}
\keyword{datasets}
