% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigos_municipios}
\alias{codigos_municipios}
\title{Codigos de los municipios de España.}
\format{
Un conjunto de datos más de 8.000 filas y 3 variables:
\describe{
  \item{codigo_provincia}{código INE de la provincia}
  \item{codigo_municipio}{código INE del municipio}
  \item{municipio}{nombre oficial más reciente del municipio}
}
}
\source{
\url{http://www.infoelectoral.mir.es/infoelectoral/min/areaDescarga.html?method=inicio}
}
\usage{
codigos_municipios
}
\description{
Este conjunto de datos contiene los códigos INE de las municipios de España con sus denominaciones más recientes (p.e: Cabrera d'Igualada aparece como Cabrera d'Anoia). Para los municipios que en algún momento han sido fusionados se mantienen sus códigos por separado junto con el del nuevo municipio creado (p.e: contiene el municipio Oza-Cesuras pero también el de Cesuras y Oza dos Ríos por separado).
}
\keyword{datasets}
