% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesas.R
\name{mesas}
\alias{mesas}
\title{Descarga datos de procesos electorales al Congreso de los Diputados, Parlamento Europeo y ayuntamientos a nivel de mesa.}
\usage{
mesas(tipo_eleccion, anno, mes)
}
\arguments{
\item{tipo_eleccion}{El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".}

\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}
}
\value{
Dataframe con los datos de voto a candidaturas por mesas.
}
\description{
Esta función descarga los datos de voto a candidaturas a nivel de mesas de las elecciones seleccionadas, los formatea, y los importa al espacio de trabajo.
}
