% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesas.R
\name{mesas}
\alias{mesas}
\title{Download data at the polling station level}
\usage{
mesas(tipo_eleccion, anno, mes)
}
\arguments{
\item{tipo_eleccion}{The type of choice you want to download. The accepted values are "congreso", "senado", "europeas" o "municipales".}

\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}
}
\value{
data.frame with the electoral results data at the polling station level.
}
\description{
`mesas()` downloads, formats and imports to the environment the electoral results data of the selected election at the polling station level.
}
\examples{
\dontrun{
  data <- mesas(tipo_eleccion = "congreso", anno = "2023", mes = "07")
  str(data)
}
}
