% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesas.R
\name{mesas}
\alias{mesas}
\title{mesas}
\usage{
mesas(url)
}
\arguments{
\item{url}{url a los datos del Ministerio del Interior de una elección determinada.}
}
\value{
Dataframe con los datos de voto a candidaturas por mesas.
}
\description{
Esta función descarga los datos de voto a candidaturas por mesas de unas elecciones, los importa al espacio de trabajo, y los formatea.
}
