% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muni.R
\name{municipios}
\alias{municipios}
\title{Descarga datos a nivel de municipio}
\usage{
municipios(tipo_eleccion, anno, mes, distritos = FALSE)
}
\arguments{
\item{tipo_eleccion}{El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".}

\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}

\item{distritos}{TRUE o FALSE, segun se quieran los resultados de los distritos de los municipios que tienen división distrital. FALSE por defecto.}
}
\value{
Dataframe con los datos de voto a candidaturas por mesas.
}
\description{
Esta función descarga los datos de voto a candidaturas a nivel de municipio de las elecciones seleccionadas, los formatea, y los importa al espacio de trabajo.
}
