% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/muni.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{Codigos de los municipios de España.}
\format{
Un conjunto de datos con 19 filas y 3 variables:
\describe{
  \item{codigo_provincia}{código INE de la provincia}
  \item{codigo_municipio}{código INE del municipio}
  \item{municipio}{nombre oficial más reciente del municipio}
}
}
\source{
\url{http://www.infoelectoral.mir.es/infoelectoral/min/areaDescarga.html?method=inicio}
}
\usage{
municipios

municipios(tipo_eleccion, anno, mes, distritos = FALSE)
}
\arguments{
\item{tipo_eleccion}{El tipo de eleccion que se quiere descargar. Los valores aceptados por ahora son "municipales" o "generales".}

\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}

\item{distritos}{TRUE o FALSE, segun se quieran los resultados de los distritos de los municipios que tienen división distrital. FALSE por defecto.}
}
\value{
Dataframe con los datos de voto a candidaturas por mesas.
}
\description{
Este conjunto de datos contiene los códigos INE de las municipios de España con sus denominaciones más recientes (p.e: Cabrera d'Igualada aparece como Cabrera d'Anoia). Para los municipios que en algún momento han sido fusionados se mantienen sus códigos por separado junto con el del nuevo municipio creado (p.e: contiene el municipio Oza-Cesuras pero también el de Cesuras y Oza dos Ríos por separado).

Esta función descarga los datos de voto a candidaturas a nivel de municipio de las elecciones seleccionadas, los formatea, y los importa al espacio de trabajo.
}
\keyword{datasets}
