% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provincias.R
\name{provincias}
\alias{provincias}
\title{Download data at the electoral constituency level (province or island)}
\usage{
provincias(tipo_eleccion, anno, mes)
}
\arguments{
\item{tipo_eleccion}{The type of choice you want to download. The accepted values are "congreso", "senado", "europeas" o "municipales".}

\item{anno}{The year of the election in YYYY format.}

\item{mes}{The month of the election in MM format.}
}
\value{
data.frame with the electoral results data at the polling station level.
}
\description{
`provincias()` downloads, formats and imports to the environment the electoral results data of the selected election at electoral constituency level (province or island).
}
