% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provincias}
\alias{provincias}
\title{Codigos INE de las provincias de España.}
\format{
Un conjunto de datos con 52 filas y 5 variables:
\describe{
  \item{codigo_ccaa}{código infoelectoral de la comunidad autónoma}
  \item{codigo_ccaa_ine}{código INE de la comunidad autónoma}
  \item{ccaa}{nombre oficial de la comunidad autónoma}
  \item{codigo_provincia}{código INE de la provincia}
  \item{provincia}{nombre oficial de la provincia}
}
}
\source{
\url{https://www.ine.es/daco/daco42/codmun/cod_provincia.htm}
}
\usage{
provincias
}
\description{
Este conjunto de datos contiene los códigos INE y los nombres de las provincias españolas.
}
\keyword{datasets}
