% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read03.R
\name{read03}
\alias{read03}
\title{Lee los ficheros de candidaturas}
\usage{
read03(file, tempd)
}
\arguments{
\item{file}{Ruta al fichero .DAT que comienza con 03 y que contienen la información de candidaturas}
}
\value{
data.frame
}
\description{
Lee los ficheros de candidaturas
}
\keyword{internal}
