% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read04.R
\name{read04}
\alias{read04}
\title{Lee los ficheros de relación de candidatos}
\usage{
read04(file, tempd)
}
\arguments{
\item{file}{Ruta al fichero .DAT que comienza con 04 y que contienen la relación de candidatos}
}
\value{
data.frame
}
\description{
Lee los ficheros de relación de candidatos
}
\keyword{internal}
