% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read05.R
\name{read05}
\alias{read05}
\title{Reads the basic municipal data files}
\usage{
read05(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 05.

 @return data.frame}
}
\description{
Reads the basic municipal data files
}
\keyword{internal}
