% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read06.R
\name{read06}
\alias{read06}
\title{Lee los ficheros de datos de candidaturas de municipios}
\usage{
read06(file, tempd)
}
\arguments{
\item{file}{Ruta al fichero .DAT que comienza con 06 y que contienen los datos de candidaturas de municipios}
}
\value{
data.frame
}
\description{
Lee los ficheros de datos de candidaturas de municipios
}
\keyword{internal}
