% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read09.R
\name{read09}
\alias{read09}
\title{Reads the basic polling station data files}
\usage{
read09(file, tempd)
}
\arguments{
\item{file}{Path to the .DAT file that begins with 09.}
}
\value{
data.frame
}
\description{
Reads the basic polling station data files
}
\keyword{internal}
