% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10.R
\name{read10}
\alias{read10}
\title{Lee los ficheros de datos de candidaturas de mesas}
\usage{
read10(file, tempd)
}
\arguments{
\item{file}{Ruta al fichero .DAT que comienza con 09 y que contienen los datos de candidaturas de mesas}
}
\value{
data.frame
}
\description{
Lee los ficheros de datos de candidaturas de mesas
}
\keyword{internal}
