% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senado_municipios.R
\name{senado_municipios}
\alias{senado_municipios}
\title{senado_municipios}
\usage{
senado_municipios(anno, mes)
}
\arguments{
\item{anno}{El año de la elección en formato YYYY. Se puede introducir como número o como texto (2015 o "2015").}

\item{mes}{El mes de la elección en formato mm. Se DEBE introducir como texto (p.e. "05" para el mes de mayo).}
}
\value{
Dataframe con los datos de candidatos.
}
\description{
Esta función descarga los datos de los candidatos al Senado a nivel de municipio.
}
